requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==1.10.19
passlib==1.7.4
python-jose==3.3.0
fastapi-crudrouter==0.8.6
psycopg2-binary==2.9.10
python-dotenv
python-multipart